after meeting with Fernando, this should be clean now
