--- assignments/life_expectancy/tests/__init__.py
@@ -1,6 +1,0 @@
-"""Test suite for the 1st assignment"""
-from pathlib import Path
-
-
-FIXTURES_DIR = Path(__file__).parent / "fixtures"
-OUTPUT_DIR = Path(__file__).parent.parent / "data"
